backend/Rag/requirements.txt
# backend/rag/requirements.txt

fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
redis[asyncio]>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0